Eliminate __wasi_lookup_t.
Instead of passing __wasi_lookup_t as a by-value struct, pass the fields
as plain scalar values. This avoids having it go through memory in the
current wasm ABI, and avoids a potential ABI break when clang
re-evaluates how big of structs it'll be willing to pass by value when
multi-value happens.