Don't duplicate signatures of syscalls
This commit adds a macro which wraps the definitions of syscalls to
generate appropriate statics and such to ensure that the actual
signatures of syscalls matches up with the signature we report to the
engine. Additionally this commit touches up syscalls to exclusively
have wasm-compatible types (i32/i64/etc) and the conversion to
u32/u8/u16/etc happens explicitly in a shim that we generate.